+++ cypress/integration/e2e/menus/emotion.spec.ts
@@ -62,12 +62,12 @@ describe('表情', () => {
         cy.get('@emotionList')
             .eq(0)
             .as('emotion')
-            .click()
+            .click({ timeout: 1000 })
             .then($el => {
                 const img = $el.find('img')
                 const src = (img.get(0) as HTMLImageElement).src
 
-                cy.get('@Editable').find('img').should('have.attr', 'src', src)
+                cy.get('@Editable').find('img', { timeout: 20000 }).should('have.attr', 'src', src)
             })
     })
 })