+++ src/config/index.ts
@@ -12,6 +12,7 @@ import imageConfig, { UploadImageHooksType } from './image'
 import textConfig from './text'
 import langConfig from './lang'
 import historyConfig from './history'
+import videoConfig from './video'
 
 // 字典类型
 export type DicType = {
@@ -45,6 +46,7 @@ export type ConfigType = {
     zIndexFullScreen: number
     showFullScreen: boolean
     showLinkImg: boolean
+    uploadImgAccept: string[]
     uploadImgServer: string
     uploadImgShowBase64: boolean
     uploadImgMaxSize: number
@@ -57,7 +59,7 @@ export type ConfigType = {
     uploadImgTimeout: number
     withCredentials: boolean
     customUploadImg: Function | null
-    customAlert: Function | null
+    customAlert: Function
 
     onCatalogChange: Function | null
 
@@ -70,6 +72,9 @@ export type ConfigType = {
     historyMaxSize: number
 
     focus: boolean
+
+    onlineVideoCheck: Function
+    onlineVideoCallback: Function
 }
 
 export type Resource = {
@@ -98,6 +103,7 @@ const defaultConfig = Object.assign(
     textConfig,
     langConfig,
     historyConfig,
+    videoConfig,
     //链接校验的配置函数
     {
         linkCheck: function (text: string, link: string): string | boolean {