+++ src/menus/img/upload-img.ts
@@ -21,24 +21,6 @@ class UploadImg {
     }
 
     /**
-     * 提示信息
-     * @param alertInfo alert info
-     * @param debugInfo debug info
-     */
-    private alert(alertInfo: string, debugInfo?: string): void {
-        const customAlert = this.editor.config.customAlert
-        if (customAlert) {
-            customAlert(alertInfo)
-        } else {
-            window.alert(alertInfo)
-        }
-
-        if (debugInfo) {
-            console.error('wangEditor: ' + debugInfo)
-        }
-    }
-
-    /**
      * 往编辑区域插入图片
      * @param src 图片地址
      */
@@ -62,8 +44,9 @@ class UploadImg {
             img = null
         }
         img.onerror = () => {
-            this.alert(
+            config.customAlert(
                 t('插入图片错误'),
+                'error',
                 `wangEditor: ${t('插入图片错误')}，${t('图片链接')} "${src}"，${t('下载链接失败')}`
             )
             img = null
@@ -155,11 +138,11 @@ class UploadImg {
         })
         // 抛出验证信息
         if (errInfos.length) {
-            this.alert(`${t('图片验证未通过')}: \n` + errInfos.join('\n'))
+            config.customAlert(`${t('图片验证未通过')}: \n` + errInfos.join('\n'), 'warning')
             return
         }
         if (resultFiles.length > maxLength) {
-            this.alert(t('一次最多上传') + maxLength + t('张图片'))
+            config.customAlert(t('一次最多上传') + maxLength + t('张图片'), 'warning')
             return
         }
 
@@ -219,7 +202,7 @@ class UploadImg {
                     if (hooks.before) return hooks.before(xhr, editor, resultFiles)
                 },
                 onTimeout: xhr => {
-                    this.alert(t('上传图片超时'))
+                    config.customAlert(t('上传图片超时'), 'error')
                     if (hooks.timeout) hooks.timeout(xhr, editor)
                 },
                 onProgress: (percent, e) => {
@@ -230,15 +213,17 @@ class UploadImg {
                     }
                 },
                 onError: xhr => {
-                    this.alert(
+                    config.customAlert(
                         t('上传图片错误'),
+                        'error',
                         `${t('上传图片错误')}，${t('服务器返回状态')}: ${xhr.status}`
                     )
                     if (hooks.error) hooks.error(xhr, editor)
                 },
                 onFail: (xhr, resultStr) => {
-                    this.alert(
+                    config.customAlert(
                         t('上传图片失败'),
+                        'error',
                         t('上传图片返回结果错误') + `，${t('返回结果')}: ` + resultStr
                     )
                     if (hooks.fail) hooks.fail(xhr, editor, resultStr)
@@ -251,8 +236,9 @@ class UploadImg {
                     }
                     if (result.errno != '0') {
                         // 返回格式不对，应该为 { errno: 0, data: [...] }
-                        this.alert(
+                        config.customAlert(
                             t('上传图片失败'),
+                            'error',
                             `${t('上传图片返回结果错误')}，${t('返回结果')} errno=${result.errno}`
                         )
                         if (hooks.fail) hooks.fail(xhr, editor, result)
@@ -271,7 +257,7 @@ class UploadImg {
             })
             if (typeof xhr === 'string') {
                 // 上传被阻止
-                this.alert(xhr)
+                config.customAlert(xhr, 'error')
             }
 
             // 阻止以下代码执行，重要！！！